Add files via upload
Exercise 1 (Crawling exercise):

A crawler is a software able to collect information from the web (and also other applications). This exercise focuses on developing a simple crawling for Wikipedia. In particular  you have to develop a software that  extracts the following information from the term “Universidad Carlos III de Madrid” in Wikipedia:

1.	Number of references to “Universidad Carlos III de Madrid” in Wikipedia.
2.	Number of students extracted from Wikipedia.

It is an open problem, i.e., you can solve it as you consider best. 
You would have to deliver:
1-	 The code you have used. 
2-	A short document (word, pdf) describing the solution you have used.